+++ packages/ApiTests/scripts/python/columns_test.py
@@ -1,8 +1,0 @@
-#name: Python Column List
-#description: column list input
-#language: python
-#input: dataframe df
-#input: column_list cols
-#output: dataframe result
-
-result = df[cols].copy()