+++ packages/ApiTests/scripts/python/single_df_test.py
@@ -1,7 +1,0 @@
-#name: Python Single Df
-#description: df input/output
-#language: python
-#input: dataframe df
-#output: dataframe result
-
-result = df