+++ packages/CVMTests/scripts/julia/calculated_column_script.jl
@@ -1,0 +1,7 @@
+#name: Julia Calc Column
+#description: calc column
+#language: julia
+#input: int x
+#output: int y
+
+y = 2 * x